update to 4

+++ src/main/java/org/b3log/symphony/processor/BreezemoonProcessor.java
@@ -20,16 +20,16 @@ package org.b3log.symphony.processor;
 import org.apache.commons.lang.StringUtils;
 import org.b3log.latke.Keys;
 import org.b3log.latke.Latkes;
-import org.b3log.latke.ioc.inject.Inject;
-import org.b3log.latke.model.User;
+import org.b3log.latke.ioc.Inject;
 import org.b3log.latke.service.LangPropsService;
+import org.b3log.latke.service.ServiceException;
 import org.b3log.latke.servlet.HTTPRequestContext;
 import org.b3log.latke.servlet.HTTPRequestMethod;
 import org.b3log.latke.servlet.annotation.After;
 import org.b3log.latke.servlet.annotation.Before;
 import org.b3log.latke.servlet.annotation.RequestProcessing;
 import org.b3log.latke.servlet.annotation.RequestProcessor;
-import org.b3log.latke.servlet.renderer.freemarker.AbstractFreeMarkerRenderer;
+import org.b3log.latke.servlet.renderer.AbstractFreeMarkerRenderer;
 import org.b3log.latke.util.Paginator;
 import org.b3log.latke.util.Requests;
 import org.b3log.symphony.model.Breezemoon;
@@ -42,7 +42,10 @@ import org.b3log.symphony.service.BreezemoonMgmtService;
 import org.b3log.symphony.service.BreezemoonQueryService;
 import org.b3log.symphony.service.DataModelService;
 import org.b3log.symphony.service.OptionQueryService;
-import org.b3log.symphony.util.*;
+import org.b3log.symphony.util.Geos;
+import org.b3log.symphony.util.Headers;
+import org.b3log.symphony.util.StatusCodes;
+import org.b3log.symphony.util.Symphonys;
 import org.json.JSONObject;
 
 import javax.servlet.http.HttpServletRequest;
@@ -62,7 +65,7 @@ import java.util.Map;
  * </ul>
  *
  * @author <a href="http://88250.b3log.org">Liang Ding</a>
- * @version 1.0.0.2, Jul 20, 2018
+ * @version 1.0.1.1, Sep 4, 2018
  * @since 2.8.0
  */
 @RequestProcessor
@@ -118,7 +121,7 @@ public class BreezemoonProcessor {
         final int pageNum = Paginator.getPage(request);
         int pageSize = Symphonys.getInt("indexArticlesCnt");
         final int avatarViewMode = (int) request.getAttribute(UserExt.USER_AVATAR_VIEW_MODE);
-        final JSONObject user = Sessions.currentUser(request);
+        final JSONObject user = (JSONObject) request.getAttribute(Common.CURRENT_USER);
         String currentUserId = null;
         if (null != user) {
             pageSize = user.optInt(UserExt.USER_LIST_PAGE_SIZE);
@@ -174,12 +177,12 @@ public class BreezemoonProcessor {
         final JSONObject breezemoon = new JSONObject();
         final String breezemoonContent = requestJSONObject.optString(Breezemoon.BREEZEMOON_CONTENT);
         breezemoon.put(Breezemoon.BREEZEMOON_CONTENT, breezemoonContent);
-        final JSONObject user = (JSONObject) request.getAttribute(User.USER);
+        final JSONObject user = (JSONObject) request.getAttribute(Common.CURRENT_USER);
         final String authorId = user.optString(Keys.OBJECT_ID);
         breezemoon.put(Breezemoon.BREEZEMOON_AUTHOR_ID, authorId);
         final String ip = Requests.getRemoteAddr(request);
         breezemoon.put(Breezemoon.BREEZEMOON_IP, ip);
-        final String ua = Headers.getHeader(request, Common.USER_AGENT);
+        final String ua = Headers.getHeader(request, Common.USER_AGENT, "");
         breezemoon.put(Breezemoon.BREEZEMOON_UA, ua);
         final JSONObject address = Geos.getAddress(ip);
         if (null != address) {
@@ -220,19 +223,26 @@ public class BreezemoonProcessor {
             return;
         }
 
-        final JSONObject breezemoon = new JSONObject();
-        breezemoon.put(Keys.OBJECT_ID, id);
-        final String breezemoonContent = requestJSONObject.optString(Breezemoon.BREEZEMOON_CONTENT);
-        breezemoon.put(Breezemoon.BREEZEMOON_CONTENT, breezemoonContent);
-        final JSONObject user = (JSONObject) request.getAttribute(User.USER);
-        final String authorId = user.optString(Keys.OBJECT_ID);
-        breezemoon.put(Breezemoon.BREEZEMOON_AUTHOR_ID, authorId);
-        final String ip = Requests.getRemoteAddr(request);
-        breezemoon.put(Breezemoon.BREEZEMOON_IP, ip);
-        final String ua = Headers.getHeader(request, Common.USER_AGENT);
-        breezemoon.put(Breezemoon.BREEZEMOON_UA, ua);
-
         try {
+            final JSONObject old = breezemoonQueryService.getBreezemoon(id);
+            if (null == old) {
+                throw new ServiceException(langPropsService.get("queryFailedLabel"));
+            }
+
+            final JSONObject user = (JSONObject) request.getAttribute(Common.CURRENT_USER);
+            if (!old.optString(Breezemoon.BREEZEMOON_AUTHOR_ID).equals(user.optString(Keys.OBJECT_ID))) {
+                throw new ServiceException(langPropsService.get("sc403Label"));
+            }
+
+            final JSONObject breezemoon = new JSONObject();
+            breezemoon.put(Keys.OBJECT_ID, id);
+            final String breezemoonContent = requestJSONObject.optString(Breezemoon.BREEZEMOON_CONTENT);
+            breezemoon.put(Breezemoon.BREEZEMOON_CONTENT, breezemoonContent);
+            final String ip = Requests.getRemoteAddr(request);
+            breezemoon.put(Breezemoon.BREEZEMOON_IP, ip);
+            final String ua = Headers.getHeader(request, Common.USER_AGENT, "");
+            breezemoon.put(Breezemoon.BREEZEMOON_UA, ua);
+
             breezemoonMgmtService.updateBreezemoon(breezemoon);
 
             context.renderJSONValue(Keys.STATUS_CODE, StatusCodes.SUCC);
@@ -256,6 +266,16 @@ public class BreezemoonProcessor {
         context.renderJSON();
 
         try {
+            final JSONObject breezemoon = breezemoonQueryService.getBreezemoon(id);
+            if (null == breezemoon) {
+                throw new ServiceException(langPropsService.get("queryFailedLabel"));
+            }
+
+            final JSONObject user = (JSONObject) request.getAttribute(Common.CURRENT_USER);
+            if (!breezemoon.optString(Breezemoon.BREEZEMOON_AUTHOR_ID).equals(user.optString(Keys.OBJECT_ID))) {
+                throw new ServiceException(langPropsService.get("sc403Label"));
+            }
+
             breezemoonMgmtService.removeBreezemoon(id);
 
             context.renderJSONValue(Keys.STATUS_CODE, StatusCodes.SUCC);
@@ -268,7 +288,7 @@ public class BreezemoonProcessor {
     private boolean isInvalid(final HTTPRequestContext context, final JSONObject requestJSONObject) {
         String breezemoonContent = requestJSONObject.optString(Breezemoon.BREEZEMOON_CONTENT);
         breezemoonContent = StringUtils.trim(breezemoonContent);
-        final int length = StringUtils.length(breezemoonContent);
+        final long length = StringUtils.length(breezemoonContent);
         if (1 > length || 512 < length) {
             context.renderMsg(langPropsService.get("breezemoonLengthLabel"));
             context.renderJSONValue(Keys.STATUS_CODE, StatusCodes.ERR);
